resolve gauri's comment

+++ google/cloud/storage/internal/async/read_range.cc
@@ -37,7 +37,7 @@ absl::optional<google::storage::v2::ReadRange> ReadRange::RangeForResume(
   std::lock_guard<std::mutex> lk(mu_);
   if (status_.has_value()) return absl::nullopt;
   if (requested_length_.has_value() && *requested_length_ >= 0 &&
-      received_bytes_ >= static_cast<std::size_t>(*requested_length_)) {
+      received_bytes_ >= *requested_length_) {
     return absl::nullopt;
   }
   range.set_read_offset(offset_);
@@ -98,10 +98,17 @@ void ReadRange::OnRead(google::storage::v2::ObjectRangeData data,
     status_ = Status{};
     CheckOverrun();
     auto result = std::move(*hash_validator_).Finish(hash_function_->Finish());
+    // GCS decompresses gzipped files on-the-fly (server-side transcoding)
+    // if decompression is not explicitly disabled by the user. In this case,
+    // the received bytes do not match the server-provided checksums (which
+    // are for the compressed object). We only bypass checksum validation if
+    // transcoding actually occurred (i.e. the received bytes do not match the
+    // compressed object size, or if the size is unknown). If they match, the
+    // file was downloaded as compressed, and we must validate the checksums
+    // to detect transit corruption.
     bool transcoded_download =
         is_transcoded_ &&
-        (!object_size_.has_value() ||
-         (received_bytes_ != static_cast<std::size_t>(*object_size_)));
+        (!object_size_.has_value() || (received_bytes_ != *object_size_));
     if (result.is_mismatch && !transcoded_download) {
       status_ = google::cloud::internal::DataLossError(
           absl::StrCat("mismatched checksums detected at the end of the "
@@ -122,12 +129,11 @@ void ReadRange::OnRead(google::storage::v2::ObjectRangeData data,
 
 void ReadRange::CheckOverrun() {
   if (requested_length_.has_value() && *requested_length_ >= 0 &&
-      received_bytes_ > static_cast<std::size_t>(*requested_length_) &&
-      !is_transcoded_ && !logged_warning_) {
+      received_bytes_ > *requested_length_ && !is_transcoded_ &&
+      !logged_warning_) {
     logged_warning_ = true;
     GCP_LOG(WARNING) << "storage: received "
-                     << (received_bytes_ -
-                         static_cast<std::size_t>(*requested_length_))
+                     << (received_bytes_ - *requested_length_)
                      << " more bytes than requested from GCS for bucket \""
                      << bucket_name_ << "\", object \"" << object_name_ << "\"";
   }
